Project1_NLP/nlp-project/requirement.txt
# Core PyTorch and transformers
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
peft>=0.4.0
accelerate>=0.20.0

# Evaluation and ML utilities
evaluate>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
jupyter>=1.0.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0